model_authentication/requirements.txt
tensorflow==2.0.0
algorithmia==1.2.1
jupyter==1.0.0
gitpython==2.1.11
matplotlib==3.0.2
urllib3==1.24.3
retry==0.9.2
numpy==1.16.1